requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
line-bot-sdk==3.6.0
pydantic-settings==2.2.1
